9949: Move report CSV into own class
